#Fix baseFragmnetでdisposeすることをやめる
